+++ domain/visit.go
@@ -1,3 +1,4 @@
+// Package domain defines the "visit" concept and the interface required by the domain/features to manage "visits".
 package domain
 
 type Visit struct {
@@ -6,7 +7,10 @@ type Visit struct {
 }
 
 // VisitsRepository is responsible for managing data related to user navigation according to the requirements provided
+//
+// Even though Store and CountUniqueVisitors can't fail when working with in-memory data structures, an error was added to the return
+// so that we can better account for future changes (e.g. using redis instead of storing everything in memory so that there's no data lost when services are shutdown)
 type VisitsRepository interface {
 	Store(visit Visit) error
-	CountUniqueVisitors(pageURL string) (int, error)
+	CountUniqueVisitors(pageURL string) (uint64, error)
 }
